backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
pgvector>=0.3.0
insightface>=0.7.3
onnxruntime>=1.18.0
hdbscan>=0.8.33
numpy>=1.26.0
Pillow>=10.0.0
python-multipart>=0.0.9
